Code refactoring and comments added
