+++ pstatus-report-sink-ktor/src/main/kotlin/gov/cdc/ocio/processingstatusapi/persistence/CosmosCollection.kt
@@ -1,0 +1,38 @@
+package gov.cdc.ocio.processingstatusapi.persistence
+
+import com.azure.cosmos.CosmosContainer
+import com.azure.cosmos.models.CosmosItemRequestOptions
+import com.azure.cosmos.models.CosmosQueryRequestOptions
+import com.azure.cosmos.models.PartitionKey
+
+/**
+ * Cosmos Collection implementation
+ *
+ * @property cosmosContainer CosmosContainer?
+ * @constructor
+ */
+class CosmosCollection(private val cosmosContainer: CosmosContainer?): Collection {
+
+    override fun <T> queryItems(query: String?, classType: Class<T>?): List<T> {
+        val items = cosmosContainer?.queryItems(
+            query, CosmosQueryRequestOptions(),
+            classType
+        )
+
+        return items?.toList() ?: listOf()
+    }
+
+    override fun <T> createItem(item: T, partitionKey: String?): Boolean {
+        val response = cosmosContainer?.createItem(
+            item,
+            PartitionKey(partitionKey),
+            CosmosItemRequestOptions()
+        )
+
+        return when (response?.statusCode) {
+            in 200..201 -> true
+            else -> false
+        }
+    }
+
+}
